# Misc performance improvements for CRAM reader