This PR tweaks the performance of the CRAM reader by the following changes:

- 16c4a00: Use lookup tables to look up rANS symbols, instead of doing binary search
- 2221b5e: Replace maps and vectors used in the rANS decoder with arrays
- 2dc6ae7: Coerce decoded quals to make `=` more efficient 
- 154426f: Replace `+` with `bit-or`

Overall, these changes make the CRAM reader roughly 4x faster than before on my environment:

```clojure
(with-open [r (reader "DRR002191.chr1.sorted.cram" {:reference ".cavia/hg38.2bit"})]
  (time (count (take 1000000 (read-alignments r)))))

;; before change
"Elapsed time: 22955.078416 msecs"

;; after change
"Elapsed time: 5079.17675 msecs"
```

Here are the profiling results before and after the change:

| before change | after change |
| :---: | :---: |
| <a href="https://github.com/chrovis/cljam/assets/27441/810ba84a-e566-43af-b76d-4486ba4d9387"><img width="500" alt="flamegraph before change" src="https://github.com/chrovis/cljam/assets/27441/810ba84a-e566-43af-b76d-4486ba4d9387"></a> | <a href="https://github.com/chrovis/cljam/assets/27441/615326c2-ffcb-47e8-8093-9a63f6436c19"><img width="500" alt="flamegraph after change" src="https://github.com/chrovis/cljam/assets/27441/615326c2-ffcb-47e8-8093-9a63f6436c19"></a> |